Cleaned up education and experience page.
Renamed the second page to education and experience (e and e). E and e is now completed. Items on e and e are hardcoded (there will not be many times when the website will need an update and is not worth the hassle of setting up database management - that's more for the blog feature). Added education, licenses and certifications, and other experiences section. Next steps: projects page, start database set up for the blog.