Blog is completed, and the website is 99% done! Whew!
This version is definitely deployable and ready to be seen by the public. The admin panels can use some CSS polishing, but a) it's not going be seen by the public, and b) the formatting of the admin panel isn't egregious. Need to upgrade to a PostgreSQL database at some point.